reorganise helper functions
