+++ frontend/cypress/support/page_objects/userRegistrationForms.js
